+++ tests/Lexer.hs
@@ -27,6 +27,8 @@
 -- ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 -- POSSIBILITY OF SUCH DAMAGE.
 
+{-# LANGUAGE TupleSections #-}
+
 module Lexer (tests) where
 
 import Control.Applicative (empty)
@@ -39,7 +41,7 @@ import Data.Char
   , isSpace
   , toLower )
 import Data.List (findIndices, isInfixOf, find)
-import Data.Maybe (listToMaybe, maybeToList, isNothing, fromJust)
+import Data.Maybe
 import Numeric (showInt, showHex, showOct, showSigned)
 
 import Test.Framework
@@ -53,10 +55,11 @@ import Text.Megaparsec.Prim
 import Text.Megaparsec.String
 import qualified Text.Megaparsec.Char as C
 
+import Prim () -- 'Arbitrary' instance for 'SourcePos'
 import Util
 
 #if !MIN_VERSION_base(4,8,0)
-import Control.Applicative ((<$>), (<*), (<*>))
+import Control.Applicative ((<$>), (<*), (<*>), (<$))
 #endif
 
 tests :: Test
@@ -64,7 +67,11 @@ tests = testGroup "Lexer"
         [ testProperty "space combinator"       prop_space
         , testProperty "symbol combinator"      prop_symbol
         , testProperty "symbol' combinator"     prop_symbol'
+        , testProperty "indentLevel"            prop_indentLevel
         , testProperty "indentGuard combinator" prop_indentGuard
+        , testProperty "nonIndented combinator" prop_nonIndented
+        , testProperty "indentBlock combinator" prop_indentBlock
+        , testProperty "indentBlock (many)"     prop_indentMany
         , testProperty "charLiteral"            prop_charLiteral
         , testProperty "integer"                prop_integer
         , testProperty "decimal"                prop_decimal
@@ -77,25 +84,25 @@ tests = testGroup "Lexer"
         , testProperty "number 2 (signed)"      prop_number_2
         , testProperty "signed"                 prop_signed ]
 
-newtype WhiteSpace = WhiteSpace
-  { getWhiteSpace :: String }
-  deriving (Show, Eq)
+-- White space
 
-instance Arbitrary WhiteSpace where
-  arbitrary = WhiteSpace . concat <$> listOf whiteUnit
+mkWhiteSpace :: Gen String
+mkWhiteSpace = concat <$> listOf whiteUnit
+  where whiteUnit = oneof [whiteChars, whiteLine, whiteBlock]
 
-newtype Symbol = Symbol
-  { getSymbol :: String }
-  deriving (Show, Eq)
+mkSymbol :: Gen String
+mkSymbol = (++) <$> symbolName <*> whiteChars
 
-instance Arbitrary Symbol where
-  arbitrary = Symbol <$> ((++) <$> symbolName <*> whiteChars)
-
-whiteUnit :: Gen String
-whiteUnit = oneof [whiteChars, whiteLine, whiteBlock]
+mkIndent :: String -> Int -> Gen String
+mkIndent x n = concat <$> sequence [spc, sym, tra, eol]
+  where spc = frequency [(5, vectorOf n itm), (1, listOf itm)]
+        tra = listOf itm
+        itm = elements " \t"
+        sym = return x
+        eol = frequency [(5, return "\n"), (1, listOf1 (return '\n'))]
 
 whiteChars :: Gen String
-whiteChars = listOf $ elements "\t\n "
+whiteChars = listOf (elements "\t\n ")
 
 whiteLine :: Gen String
 whiteLine = commentOut <$> arbitrary `suchThat` goodEnough
@@ -118,52 +125,124 @@ sc = space (void C.spaceChar) l b
 sc' :: Parser ()
 sc' = space (void $ C.oneOf " \t") empty empty
 
-prop_space :: WhiteSpace -> Property
-prop_space w = checkParser p r s
+prop_space :: Property
+prop_space = forAll mkWhiteSpace (checkParser p r)
   where p = sc
         r = Right ()
-        s = getWhiteSpace w
 
-prop_symbol :: Symbol -> Maybe Char -> Property
-prop_symbol = parseSymbol (symbol sc) id
+prop_symbol :: Maybe Char -> Property
+prop_symbol t = forAll mkSymbol $ \s ->
+  parseSymbol (symbol sc) id s t
 
-prop_symbol' :: Symbol -> Maybe Char -> Property
-prop_symbol' = parseSymbol (symbol' sc) (fmap toLower)
+prop_symbol' :: Maybe Char -> Property
+prop_symbol' t = forAll mkSymbol $ \s ->
+  parseSymbol (symbol' sc) (fmap toLower) s t
 
-parseSymbol :: (String -> Parser String) -> (String -> String)
-            -> Symbol -> Maybe Char -> Property
+parseSymbol
+  :: (String -> Parser String)
+  -> (String -> String)
+  -> String
+  -> Maybe Char
+  -> Property
 parseSymbol p' f s' t = checkParser p r s
   where p = p' (f g)
         r | g == s || isSpace (last s) = Right g
           | otherwise = posErr (length s - 1) s [uneCh (last s), exEof]
         g = takeWhile (not . isSpace) s
-        s = getSymbol s' ++ maybeToList t
+        s = s' ++ maybeToList t
 
-newtype IndLine = IndLine
-  { getIndLine :: String }
-  deriving (Show, Eq)
+-- Indentation
 
-instance Arbitrary IndLine where
-  arbitrary = IndLine . concat <$> sequence [spc, sym, spc, eol]
-    where spc = listOf (elements " \t")
-          sym = return "xxx"
-          eol = return "\n"
+prop_indentLevel :: SourcePos -> Property
+prop_indentLevel pos = p /=\ sourceColumn pos
+  where p = setPosition pos >> indentLevel
 
-prop_indentGuard :: IndLine -> IndLine -> IndLine -> Property
-prop_indentGuard l0 l1 l2 = checkParser p r s
-  where p  = ip (> 1) >>= \x -> sp >> ip (== x) >> sp >> ip (> x) >> sp
-        ip = indentGuard sc'
-        sp = void $ symbol sc' "xxx" <* C.eol
-        r | f' l0 <= 1     = posErr 0 s msg'
-          | f' l1 /= f' l0 = posErr (f l1 + g [l0]) s msg'
-          | f' l2 <= f' l0 = posErr (f l2 + g [l0, l1]) s msg'
+prop_indentGuard :: NonNegative Int -> Property
+prop_indentGuard n =
+  forAll ((,,) <$> mki <*> mki <*> mki) $ \(l0,l1,l2) ->
+    let r | getCol l0 <= 1         = posErr 0 s ii
+          | getCol l1 /= getCol l0 = posErr (getIndent l1 + g 1) s ii
+          | getCol l2 <= getCol l0 = posErr (getIndent l2 + g 2) s ii
           | otherwise = Right ()
-        msg' = [msg "incorrect indentation"]
-        f    = length . takeWhile isSpace . getIndLine
-        f' x = sourceColumn $ updatePosString defaultTabWidth (initialPos "") $
-               take (f x) (getIndLine x)
-        g xs = sum $ length . getIndLine <$> xs
-        s    = concat $ getIndLine <$> [l0, l1, l2]
+        fragments = [l0,l1,l2]
+        g x = sum (length <$> take x fragments)
+        s = concat fragments
+    in checkParser p r s
+  where mki = mkIndent sbla (getNonNegative n)
+        p  = ip (> 1) >>= \x -> sp >> ip (== x) >> sp >> ip (> x) >> sp >> sc
+        ip = indentGuard sc
+        sp = void (symbol sc' sbla <* C.eol)
+
+prop_nonIndented :: Property
+prop_nonIndented = forAll (mkIndent sbla 0) $ \s ->
+  let i = getIndent s
+      r | i == 0    = Right sbla
+        | otherwise = posErr i s ii
+  in checkParser p r s
+  where p = nonIndented sc (symbol sc sbla)
+
+prop_indentBlock :: Maybe (Positive Int) -> Property
+prop_indentBlock mn' = forAll mkBlock $ \(l0,l1,l2,l3,l4) ->
+  let r | getCol l1 <= getCol l0 =
+          posErr (getIndent l1 + g 1) s [uneCh (head sblb), exEof]
+        | isJust mn && getCol l1 /= ib =
+          posErr (getIndent l1 + g 1) s ii
+        | getCol l2 <= getCol l1 =
+          posErr (getIndent l2 + g 2) s ii
+        | getCol l3 == getCol l2 =
+          posErr (getIndent l3 + g 3) s [uneCh (head sblb), exStr sblc]
+        | getCol l3 <= getCol l0 =
+          posErr (getIndent l3 + g 3) s [uneCh (head sblb), exEof]
+        | getCol l3 /= getCol l1 =
+          posErr (getIndent l3 + g 3) s ii
+        | getCol l4 <= getCol l3 =
+          posErr (getIndent l4 + g 4) s ii
+        | otherwise = Right (sbla, [(sblb, [sblc]), (sblb, [sblc])])
+      fragments = [l0,l1,l2,l3,l4]
+      g x = sum (length <$> take x fragments)
+      s = concat fragments
+  in checkParser p r s
+  where mkBlock = do
+          l0 <- mkIndent sbla 0
+          l1 <- mkIndent sblb ib
+          l2 <- mkIndent sblc (ib + 2)
+          l3 <- mkIndent sblb ib
+          l4 <- mkIndent sblc (ib + 2)
+          return (l0,l1,l2,l3,l4)
+        p = lvla
+        lvla = indentBlock sc $ IndentMany mn      (l sbla) lvlb <$ b sbla
+        lvlb = indentBlock sc $ IndentSome Nothing (l sblb) lvlc <$ b sblb
+        lvlc = indentBlock sc $ IndentNone                  sblc <$ b sblc
+        b    = symbol sc'
+        l x  = return . (x,)
+        mn   = getPositive <$> mn'
+        ib   = fromMaybe 2 mn
+
+prop_indentMany :: Property
+prop_indentMany = forAll (mkIndent "xxx" 0) (checkParser p r)
+  where r = Right (sbla, [])
+        p = lvla
+        lvla = indentBlock sc $ IndentMany Nothing (l sbla) lvlb <$ b sbla
+        lvlb = b sblb
+        b    = symbol sc'
+        l x  = return . (x,)
+
+getIndent :: String -> Int
+getIndent = length . takeWhile isSpace
+
+getCol :: String -> Int
+getCol x = sourceColumn $
+  updatePosString defaultTabWidth (initialPos "") $ take (getIndent x) x
+
+sbla, sblb, sblc :: String
+sbla = "xxx"
+sblb = "yyy"
+sblc = "zzz"
+
+ii :: [Message]
+ii = [msg "incorrect indentation"]
+
+-- Character and string literals
 
 prop_charLiteral :: String -> Bool -> Property
 prop_charLiteral t i = checkParser charLiteral r s
@@ -175,6 +254,8 @@ prop_charLiteral t i = checkParser charLiteral r s
           | otherwise   = posErr l s [uneCh (head g), exEof]
         l = length s - length g
         s = if null t || i then t else showLitChar (head t) (tail t)
+
+-- Numbers
 
 prop_integer :: NonNegative Integer -> Int -> Property
 prop_integer n' i = checkParser integer r s
@@ -227,11 +308,11 @@ prop_signed n i plus = checkParser p r s
   where p = signed (hidden C.space) integer
         r | i > length z = Right n
           | otherwise = posErr i s $ uneCh '?' :
-                        (if i <= 0 then [exCh '+', exCh '-'] else []) ++
-                        [exSpec $ if isNothing . find isDigit $ take i s
-                                  then "integer"
-                                  else "rest of integer"]
-                        ++ [exEof | i > head (findIndices isDigit s)]
+            (if i <= 0 then [exCh '+', exCh '-'] else []) ++
+            [exSpec $ if isNothing . find isDigit $ take i s
+                      then "integer"
+                      else "rest of integer"] ++
+            [exEof | i > head (findIndices isDigit s)]
         z = let bar = showSigned showInt 0 n ""
             in if n < 0 || plus then bar else '+' : bar
         s = if i <= length z then take i z ++ "?" ++ drop i z else z
@@ -243,9 +324,9 @@ quasiCorrupted n' i shower l = (r, s)
   where n = getNonNegative n'
         r | i > length z = Right n
           | otherwise    = posErr i s $ uneCh '?' :
-                           [ exEof | i > 0 ] ++
-                           [if i <= 0 || null l
-                            then exSpec l
-                            else exSpec $ "rest of " ++ l]
+            [ exEof | i > 0 ] ++
+            [if i <= 0 || null l
+             then exSpec l
+             else exSpec $ "rest of " ++ l]
         z = shower n ""
         s = if i <= length z then take i z ++ "?" ++ drop i z else z